refactor: move OllamaJudge to helpers module
Move OllamaJudge and related code from test_ollama_judge.py to helpers.py
for better separation of concerns. This keeps conftest.py focused on pytest
configuration while helpers.py contains reusable test helper logic.
